♻️ refactor: 좋아요 버튼 컴포넌트 분리 (#76)
