+++ src/components/ProductCard.tsx
@@ -1,35 +1,14 @@
 import { GetProductType, LikedProductType } from 'types/product';
 import { useNavigate } from 'react-router-dom';
-import { useAuthContext } from 'context/AuthContext';
-import { AuthContextType } from 'types/auth';
-import HeartSvg from 'assets/svg/HeartSvg';
-import { useModalStore } from 'store/modal';
-import LoginRequestModal from './LoginRequestModal';
-import { useProductLikeToggle } from 'hooks/useProductLikeToggle';
+import LikeButton from './LikeButton';
 
 type Props = {
   product: GetProductType | LikedProductType;
 };
 
 const ProductCard = ({ product }: Props) => {
-  const { id, name, mainImg, price, heartCount, isLiked } = product;
-  const { user } = useAuthContext() as AuthContextType;
-  const { mutate: toggleLikeMutate } = useProductLikeToggle(
-    id,
-    isLiked,
-    user?.uid,
-  );
-  const { openModal } = useModalStore();
+  const { id, name, mainImg, price, isLiked, heartCount } = product;
   const navigate = useNavigate();
-
-  const handleLike = (e: React.MouseEvent<HTMLButtonElement>) => {
-    e.stopPropagation();
-    if (!user?.uid) {
-      openModal(<LoginRequestModal />);
-      return;
-    }
-    toggleLikeMutate({ uid: user?.uid, productId: id });
-  };
 
   return (
     <li className='cursor-pointer' onClick={() => navigate(`/products/${id}`)}>
@@ -43,16 +22,8 @@ const ProductCard = ({ product }: Props) => {
       </div>
       <div className='mt-2 px-2'>
         <h3 className='truncate'>{name}</h3>
-        <p className='font-semibold text-price'>{`${price.toLocaleString()}원`}</p>
-        <button
-          className='flex items-center'
-          onClick={handleLike}
-          disabled={user?.isAdmin}
-          title={user?.isAdmin ? '좋아요는 일반 회원만 가능합니다.' : ''}
-        >
-          <HeartSvg isLiked={isLiked} />
-          <p className='m-1'>{heartCount || 0}</p>
-        </button>
+        <p className='font-semibold text-price mb-1'>{`${price.toLocaleString()}원`}</p>
+        <LikeButton isLiked={isLiked} id={id} heartCount={heartCount} />
       </div>
     </li>
   );
